+++ CHANGES.md
@@ -6,6 +6,8 @@
 - when using "split tunnel" configuration, also set a static route to the VPN 
   server through the client's normal gateway to avoid problems when pushing 
   routes that contain the VPN server's public IP address
+- remove compression framing support
+- remove tls-auth support
 
 ## 1.1.2 (2018-11-09)
 - add support for `blockLan` to block traffic to local LAN network when